added: edit subject view

+++ resources/components/EditSubject.tsx
@@ -16,37 +16,38 @@ const groupNames = ['1EF-DI', '2EF-DI', '3EF-DI', '1XF-XD', '2KE-KW']
 
 export const EditSubject = (props: Props) => {
   const { errors, register, handleSubmit, reset, formState, setValue } = useForm({ mode: 'onSubmit' })
+  const [subject, setSubject] = useState<TSubject | null | undefined>(null)
   const [groups, setGroups] = useState<string | null>(null)
+  const [submitType, setSubmitType] = useState<'delete' | 'save' | null>(null)
   const [feedback, setFeedback] = useState<TMessage>({ text: '' })
 
   const classId = props.match.params.id
 
-  useEffect(() => {
-    request<TSubject>(`subjects/${classId}`).then(({ name, description, group, subgroup }) => {
-      setValue('name', name)
-      setValue('description', description)
-      setValue('group', '3EF-DI')
-      setValue('subgroup', '4')
-    })
-    // request<string[]>('groups').then(Object.values).then(setGroups)
-  }, [])
+  useEffect(() => void (async () => {
+    // await request<string[]>('groups').then(Object.values).then(setGroups)
+
+    const subject = await request<TSubject>(`subjects/${classId}`)
+    const { name, description, group, subgroup } = subject
+    const formData = { name, description, group, subgroup }
+
+    setSubject(subject)
+    Object.entries(formData).forEach(([k, v]) => setValue(k, v))
+  })(), [])
 
   const onSubmit = (payload: Record<string, string>) => {
-    return request<string>('subjects', { method: 'post', body: JSON.stringify(payload) })
+    if (submitType === 'delete') {
+      const msg = `Are you sure that you want to delete the class '${subject?.name}'?`
+      return confirm(msg) && request<void>(`/subjects/${classId}`, { method: 'delete' }).then(() => {
+        setSubject(void 0)
+        setFeedback({ variant: 'success', text: `You have successfully deleted '${subject?.name}'!`} )
+      }).catch(() => setFeedback({ variant: 'error', text: 'An error has occurred. Please try again later'} ))
+    }
+
+    return request<string>(`subjects/${classId}`, { method: 'patch', body: JSON.stringify(payload) })
       .then(() => {
-        reset({ name: '', group: null, subgroup: null, description: '' })
-        setFeedback({ variant: 'success', text: 'You have successfully created a new class!'} )
+        setFeedback({ variant: 'success', text: 'You have successfully updated the information!'} )
       })
       .catch(() => setFeedback({ variant: 'error', text: 'An error has occurred. Please try again later'} ))
-  }
-
-  const onDelete = () => {
-    if (!confirm('Are you sure?')) return
-
-    request<void>(`/subjects/${classId}`, { method: 'delete' }).then(() => {
-      reset({ name: '', group: null, subgroup: null, description: '' })
-      setFeedback({ variant: 'success', text: 'You have successfully deleted this class!'} )
-    }).catch(() => setFeedback({ variant: 'error', text: 'An error has occurred. Please try again later'} ))
   }
 
   const validateName = (input: string) => {
@@ -69,6 +70,8 @@ export const EditSubject = (props: Props) => {
     if (input.indexOf('\\') !== -1) return 'Description contains a forbidden character!'
   }
 
+  const disabled = () => subject === void 0 ? ' opacity-50 pointer-events-none' : ''
+
   return (
     <div className='stack'>
       <div className='hstack mb-5 justify-between'>
@@ -76,31 +79,34 @@ export const EditSubject = (props: Props) => {
         <Link className='self-center border-current border-1 px-3 py-1 cursor-pointer hover:text-white hover:bg-black dark:hover:text-black dark:hover:bg-gray-200' to={`/classes/${classId}`}>return</Link>
       </div>
 
-      {/* {groups === null && <Loader />} */}
+      {(subject === null || groups === null) && <Loader />}
 
-      <form noValidate className={`grid gap-10 gap-y-5 grid-cols-2 w-full max-w-screen-lg mx-auto${groups !== null ? ' opacity-50 pointer-events-none' : ''}`} onSubmit={handleSubmit(onSubmit)}>
+      <form noValidate className={`grid gap-10 gap-y-5 grid-cols-2 w-full max-w-screen-lg mx-auto${(subject === null || groups === null) ? ' opacity-50 pointer-events-none' : ''}`} onSubmit={handleSubmit(onSubmit)}>
         <InputText
-          className='col-span-full' name='name' variant='underlined' label='class name' placeholder='class name'
+          className={`col-span-full${disabled()}`} name='name' variant='underlined' label='class name' placeholder='class name'
           required maxLength={64} ref={register({ validate: validateName })} error={errors?.name?.message}
         />
-        <Select name='group' placeholder='group name' error={errors?.group?.message} ref={register({ validate: validateGroup })}>
+        <Select className={disabled()} name='group' placeholder='group name' error={errors?.group?.message} ref={register({ validate: validateGroup })}>
           {groupNames.map((v, i) => <option key={i} value={v}>{v}</option>)}
         </Select>
-        <Select name='subgroup' placeholder='subgroup' error={errors?.subgroup?.message} ref={register({ validate: validateSubgroup })}>
+        <Select className={disabled()} name='subgroup' placeholder='subgroup' error={errors?.subgroup?.message} ref={register({ validate: validateSubgroup })}>
           {[...new Array(16)].map((v, i) => <option key={i} value={i+1}>L{i+1}</option>)}
         </Select>
         <InputText
-          className='col-span-full' name='description' variant='underlined' label='class description' placeholder='class description'
+          className={`col-span-full${disabled()}`} name='description' variant='underlined' label='class description' placeholder='class description'
           ref={register({ validate: validateDescription })} error={errors?.description?.message}
         />
         <Message ctx={feedback} className='col-span-full' onClose={() => setFeedback({ text: '' })} />
         <input
-          type='button' value='delete' onClick={onDelete}
+          type='submit' value='delete' onClick={() => setSubmitType('delete')} disabled={formState.isSubmitting || subject === void 0}
           className={`col-auto mr-auto mt-2 px-10 border-current border-1 py-1 cursor-pointer bg-transparent \
-hover:text-white dark:hover:text-black focus:outline-none text-red-500 hover:bg-red-500`}
+hover:text-white dark:hover:text-black focus:outline-none text-red-500 hover:bg-red-500 disabled:cursor-wait disabled:opacity-20`}
         />
-        <input type='submit' value='save' disabled={formState.isSubmitting} className={`col-auto ml-auto mt-2 px-10 border-current border-1 py-1 cursor-pointer bg-transparent \
-hover:text-white hover:bg-black dark:hover:text-black dark:hover:bg-gray-200 focus:outline-none disabled:cursor-wait disabled:opacity-20`} />
+        <input
+          type='submit' value='save' disabled={formState.isSubmitting || subject === void 0} onClick={() => setSubmitType('save')}
+          className={`col-auto ml-auto mt-2 px-10 border-current border-1 py-1 cursor-pointer bg-transparent \
+hover:text-white hover:bg-black dark:hover:text-black dark:hover:bg-gray-200 focus:outline-none disabled:cursor-wait disabled:opacity-20`}
+        />
       </form>
     </div>
   )
